requirements.txt
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
pandas==2.2.2
numpy==1.26.4
websocket-client==1.8.0
pytz
scikit-learn
scipy
matplotlib
seaborn
websocket
aiohttp
pydantic
websockets
openai
flask
scipy
statsmodels
